Move ZeroCopyBuf to its own module
